Add category "IntegrationTests" to split tests required for CI

+++ mico-core/src/test/java/io/github/ust/mico/core/ImageBuilderIntegrationTests.java
@@ -6,17 +6,14 @@ import io.github.ust.mico.core.build.ImageBuilderConfig;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
+import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.util.ResourceUtils;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CountDownLatch;
@@ -25,6 +22,7 @@ import java.util.concurrent.TimeUnit;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
+@Category(IntegrationTests.class)
 @Slf4j
 @SpringBootTest
 @RunWith(SpringJUnit4ClassRunner.class)
@@ -44,20 +42,29 @@ public class ImageBuilderIntegrationTests {
 
     private CountDownLatch lock = new CountDownLatch(1);
 
+    private static boolean setUpFinished = false;
+    private static boolean tearDownFinished = false;
+
     @Before
     public void setUp() {
-        //Namespaces
-        cluster.createNamespace(config.getNamespaceName());
+        if (!setUpFinished) {
+            //Namespaces
+            cluster.createNamespace(config.getNamespaceName());
 
-        // TODO Why works only the `default` namespace?!
-        // Override namespace
-        //imageBuilderConfig.setBuildExecutionNamespace(config.getNamespaceName());
+            // TODO Why works only the `default` namespace?!
+            // Override namespace
+            //imageBuilderConfig.setBuildExecutionNamespace(config.getNamespaceName());
+            setUpFinished = true;
+        }
     }
 
     @After
     public void tearDown() {
-        //Namespaces
-        cluster.deleteNamespace(config.getNamespaceName());
+        if (!tearDownFinished) {
+            //Namespaces
+            cluster.deleteNamespace(config.getNamespaceName());
+            tearDownFinished = true;
+        }
     }
 
     @Test
@@ -75,21 +82,12 @@ public class ImageBuilderIntegrationTests {
         assertNotNull("No Build CRD defined", buildCRD);
     }
 
-    @Test
-    public void createBuildObjectWithYaml() throws IOException {
-
-        File file = ResourceUtils.getFile("classpath:hello-build.yaml");
-        InputStream yaml = new FileInputStream(file);
-        System.out.println("Start Build");
-        imageBuilder.createBuildWithYaml(yaml, config.getNamespaceName());
-        System.out.println("Finished Build");
-    }
-
     @Test(expected = Exception.class)
     public void withoutInitializingAnErrorIsThrown() throws Exception {
         imageBuilder.build("service-name", "1.0.0", "Dockerfile", "https://github.com/dgageot/hello.git", "master");
     }
 
+    @Ignore
     @Test
     public void createHelloBuildImage() throws Exception {
